+++ frontend/webpack/_base.js
@@ -1,0 +1,27 @@
+/**
+ * Base configuration to share with all concrete configurations
+ */
+import MiniCssExtractPlugin from "mini-css-extract-plugin";
+
+export default {
+    // Modules rules
+    module: {
+        rules: [
+            // Babel ES6 inspection watch for every JS sources changes
+            {
+                test: /\.js$/,
+                exclude: /node_modules/,
+                use: {
+                    loader: "babel-loader",
+                    options: {
+                        presets: ["@babel/preset-env"]
+                    },
+                }
+            },
+            {
+                test: /\.css$/,
+                use: [MiniCssExtractPlugin.loader, "css-loader"]
+            }
+        ]
+    },
+};